organized helper functions better
