+++ contracts/mesh-consumer/src/testing/utils/mod.rs
@@ -1,3 +1,4 @@
+pub mod executes;
 pub mod helpers;
 
 pub mod setup;
